requirements/base.txt
argon2-cffi==21.3.0
celery==5.2.7
cloudinary==1.29.0
coverage==6.4.4
Django==4.1.1
django-debug-toolbar==3.6.0
django-environ==0.8.1
django-filter==21.1
django-money==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2-binary==2.9.3
redis==4.3.4
google-api-python-client==2.65.0
facebook_business==15.0.0
twitter==1.19.6